+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * A potential crash when creating the `HeifDepthRepresentationInfo`.
+* A parameter validation issue when reading zero bytes.
 
 ## v2.1.0 - 2021-05-19